requirements.txt
# Core
python>=3.9

# HTTP client for Ollama
requests>=2.31.0

# Optional: environment variable loading
python-dotenv>=1.0.1

# From traditional detector to support PDF/DOCX and Streamlit UI
pdfplumber>=0.11.4
python-docx>=1.1.2
streamlit==1.32.2
